Properly validate fields when resetting password
The logic for validating secure storage passwords is designed
for when a user creates a new password and not when they reset
it.  This causes the dialog to enable the "Reset" button prematurely.

This patch adds the necessary checks for when mResetPassword is true

Change-Id: I6d52865b2316ff6eeffeb5eaf61e787f6222e9f7
(cherry picked from commit 215e74651e478c148a92695d86e36d100711ab63)