CMFileManager: Fix ANR when press back key in File Manager
When back key pressed, it path in history was checked to make sure it
was still available, but the getFileInfo command cost very long time
if the folder contained a lot of files, that blocked the UI thread and
caused ANR occurred.

Check the folder available by File.exists() will reduce much time.

Change-Id: Idbf3c1b4f2cd6aff421725bacc92d827ee29c32f
CRs-Fixed: 928717